feat(variables): same variables for timers as in commands
