+++ systems/timers.js
@@ -1,6 +1,8 @@
 const _ = require('lodash')
 const { say } = require('./customCommands')
 const { io } = require("../libs/panel")
+const commons = require('../libs/commons')
+const variables = require('../systems/variables')
 
 class Timers {
   constructor() {
@@ -19,11 +21,18 @@ class Timers {
     for (let timer of timers) {
       if ((new Date().getTime() - timer.triggertimestamp) > (timer.interval * 1000)) {
         if (!global.tmi.uptime) return
-        await say(timer.responses[timer.last])
+        let response = await variables.prepareMessage(timer.responses[timer.last], { 'display-name': process.env.TWITCH_BOTUSERNAME })
+        await say(response)
         await global.db('systems.timers').where('name', timer.name).update({ last: ++timer.last % timer.responses.length, triggertimestamp: new Date().getTime() })
       }
     }
     this.timeout = setTimeout(() => this.check(), 10000)
+  }
+  async prepareResponse(response) {
+    if (response.includes('(api|')) {
+      response = await commons.parseMessageApi(response, { 'display-name': process.env.TWITCH_BOTUSERNAME })
+    }
+    return response
   }
   async sockets() {
     let self = this
